+++ src/contentScript/shortcutSnippetMenu/ShortcutSnippetMenu.scss
