+++ src/contentScript/shortcutSnippetMenu/shortcutSnippetMenuSlice.ts
