+++ src/contentScript/shortcutSnippetMenu/shortcutSnippetUtils.ts
