+++ src/tsconfig.strictNullChecks.json
@@ -296,16 +296,16 @@
     "./components/walkthroughModal/showWalkthroughModal.ts",
     "./contentScript/activationConstants.ts",
     "./contentScript/browserActionInstantHandler.ts",
-    "./contentScript/shortcutSnippetMenu/CommandPopover.stories.tsx",
-    "./contentScript/shortcutSnippetMenu/CommandPopover.test.tsx",
-    "./contentScript/shortcutSnippetMenu/CommandPopover.tsx",
-    "./contentScript/shortcutSnippetMenu/CommandRegistry.ts",
-    "./contentScript/shortcutSnippetMenu/commandController.tsx",
-    "./contentScript/shortcutSnippetMenu/commandPopoverSlice.ts",
-    "./contentScript/shortcutSnippetMenu/commandPopoverSlice.test.ts",
-    "./contentScript/shortcutSnippetMenu/commandUtils.ts",
-    "./contentScript/shortcutSnippetMenu/commandUtils.test.ts",
-    "./contentScript/shortcutSnippetMenu/useCommandRegistry.ts",
+    "./contentScript/shortcutSnippetMenu/ShortcutSnippetMenu.stories.tsx",
+    "./contentScript/shortcutSnippetMenu/ShortcutSnippetMenu.test.tsx",
+    "./contentScript/shortcutSnippetMenu/ShortcutSnippetMenu.tsx",
+    "./contentScript/shortcutSnippetMenu/ShortcutSnippetRegistry.ts",
+    "./contentScript/shortcutSnippetMenu/shortcutSnippetMenuController.tsx",
+    "./contentScript/shortcutSnippetMenu/shortcutSnippetMenuSlice.ts",
+    "./contentScript/shortcutSnippetMenu/shortcutSnippetMenuSlice.test.ts",
+    "./contentScript/shortcutSnippetMenu/shortcutSnippetUtils.ts",
+    "./contentScript/shortcutSnippetMenu/shortcutSnippetUtils.test.ts",
+    "./contentScript/shortcutSnippetMenu/useShortcutSnippetRegistry.ts",
     "./contentScript/shortcutSnippetMenu/useKeyboardQuery.ts",
     "./contentScript/context.ts",
     "./contentScript/contextMenus.ts",
